Fix %u format specifier left mismatched by the txn-id-64bit merge
Merging origin/main (HarperFast/rocksdb-js#853, which widened
TransactionHandle::id to uint64_t) brought every other DEBUG_LOG call site in
this file up to %llu + an explicit cast; this branch's own admission-refusal
log line, added independently, was the one site the merge could not touch and
was left passing a uint64_t through a %u specifier.

+++ src/binding/transaction/transaction.cpp
@@ -438,8 +438,8 @@ static void executeLogWork(TransactionCommitState* state) {
 		DEBUG_LOG("%p Transaction::Commit ERROR: Called with dbHandle not opened\n", txnHandle.get());
 		state->status = rocksdb::Status::Aborted("Database closed during transaction commit operation");
 	} else if (rocksdb::Status admission = admitCommit(state, txnHandle); !admission.ok()) {
-		DEBUG_LOG("%p Transaction::Commit refused at admission for transaction %u: %s\n",
-			txnHandle.get(), txnHandle->id, admission.ToString().c_str());
+		DEBUG_LOG("%p Transaction::Commit refused at admission for transaction %llu: %s\n",
+			txnHandle.get(), (unsigned long long)txnHandle->id, admission.ToString().c_str());
 		state->status = admission;
 	} else if (txnHandle->logEntryBatch) {
 		DEBUG_LOG("%p Transaction::Commit Committing log entries for transaction %llu\n",
